requirements.txt
fastapi
uvicorn
pydantic
soundfile
numpy
pydub
emoji

torch
torchaudio
transformers>=4.33,<=4.46.2

coqui-tts
